backend/requirements.txt
# Django Core
setuptools==69.0.3
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# MySQL
mysqlclient==2.2.1

# WebSocket & Video
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# AI - FREE
google-generativeai==0.3.2

# TTS - Install separately to avoid conflicts
TTS

# Core dependencies for TTS (Python 3.10 compatible)
numpy
scipy
numba

# Audio Processing
librosa
soundfile
pydub

# Video Processing
opencv-python
Pillow
imageio
imageio-ffmpeg

# Utilities
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
